Organizing imports, moving prop types to single component
