+++ src/components/pages/Home/About/Form/PaymentDiv.tsx
@@ -1,3 +1,4 @@
+//Image imports
 import checkmark from "../../../../../assets/index/icons/icon-checkmark.svg"
 
 const PaymentDiv = () => {
