+++ src/components/pages/Home/Blog/BlogHighlights.tsx
@@ -1,3 +1,4 @@
+//react imports
 import { NavLink } from "react-router-dom"
 //component imports
 import HighlightArticle from "./HighlightArticle"
@@ -5,7 +6,6 @@ import HighlightArticle from "./HighlightArticle"
 import articleImgOne from "../../../../assets/index/photos/photo-wfh-thought-catalog.jpg"
 import articleImgTwo from "../../../../assets/index/photos/photo-businessman-taylor-grote.jpg"
 import articleImgThree from "../../../../assets/index/photos/photo-laptop-corinne-kutz.jpg"
-
 
 const BlogHighlights = () => {
   return (
