+++ src/components/pages/Home/Blog/HighlightArticle.tsx
@@ -3,19 +3,16 @@ import { NavLink } from "react-router-dom"
 //ionicons
 import { IonIcon } from "@ionic/react"
 import { arrowForwardOutline } from "ionicons/icons"
+//types
+import { HighlightArticleProps } from "../../../../types/propTypes"
 
-type Props = {
-    img: string,
-    alt: string
-}
-
-const HighlightArticle = ({ img, alt }: Props) => {
+const HighlightArticle = ({ img, alt }: HighlightArticleProps) => {
     return (
         <article className="lg:w-1/4">
             <NavLink to="/" className="flex flex-col gap-4 p-4 transition-all hover:bg-jobi-green-light hover:-translate-y-4 focus:bg-jobi-green-light focus:-translate-y-4 group sm:flex-row lg:flex-col">
                 <span className="sr-only">Click to read this blog post</span>
                 <div className="sm:w-2/3 lg:w-full">
-                    <img src={img} alt={alt} title={alt} />
+                    <img src={img} alt={alt} title="Blog image" />
                 </div>
                 <div className="flex flex-col justify-evenly gap-2">
                     <p className="text-xs font-gorditaLight">22 July, 2022</p>
