+++ src/types/propTypes.ts
@@ -1,5 +1,11 @@
 import { ReactNode } from "react"
 
+export type AccordionItemProps = {
+    children: ReactNode,
+    heading: string,
+    index: number
+}
+
 type ButtonColorClasses = 'btn-green' | 'btn-outline' | 'btn-lime' | 'btn-blue' | 'btn-cream' | 'btn-pink' | 'btn-lightgreen' | 'btn-lavender' | 'btn-orange'
 
 export interface ButtonProps {
@@ -13,16 +19,30 @@ export interface AnchorProps extends ButtonProps {
     url: string
 }
 
+export type HighlightArticleProps = {
+    img: string,
+    alt: string
+}
+
 export type ListItemLinkProps = {
     url: string,
     text: string,
     extraClasses?: string
 }
-// HowItWorks component step-by-step cards
+
 export type StepCardProps = {
     src: string,
     text: string,
     url: string,
     linkText: string,
     stepNumber: string
+}
+
+
+export type TestimonialCarouselProps = {
+    heading: string,
+    text: string,
+    name: string,
+    location: string,
+    img: string
 }
